backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
python-dotenv
openai
gigachat
sentence-transformers
numpy
networkx
pdfplumber
aiofiles
requests
pydantic-settings
faiss-cpu
sqlalchemy
psycopg2-binary
python-jose[cryptography]
bcrypt